use header to find comment
In https://github.com/iFixit/contextual-qa-checklist-action/pull/3
we removed the default string for the footer and set footer to not be
required.

This introduced a bug because the default for the footer is an empty
string. As a result, when we looked for the first instance of a comment
that CONTAINED an empty string, it would always return the first
comment in the list, thus it would always delete any comment.

We will now search for a comment that includes the header since that is
required.